#233 Restructure file
