+++ MindSpace.Application/Features/Appointments/Commands/HandleWebhook/HandleWebhookCommandHandler.cs
@@ -1,7 +1,7 @@
 ﻿using AutoMapper;
 using MediatR;
 using Microsoft.Extensions.Logging;
-using MindSpace.Application.BackgroundJobs;
+using MindSpace.Application.BackgroundJobs.MeetingRooms;
 using MindSpace.Application.Commons.Constants;
 using MindSpace.Application.DTOs.Notifications;
 using MindSpace.Application.Interfaces.Repos;
